refactor: Move SchemaValidator out of dml handlers
Leave the implementation of DmlHandler on SchemaValidator where it is,
but create a SchemaValidator and start a validate method (that will
probably become "upsert") to be shared between write and schema
operations.
