+++ ioxd_router/src/lib.rs
@@ -51,7 +51,7 @@ use object_store::DynObjectStore;
 use router::{
     dml_handlers::{
         lazy_connector::LazyConnector, DmlHandler, DmlHandlerChainExt, FanOutAdaptor,
-        InstrumentationDecorator, Partitioner, RetentionValidator, RpcWrite, SchemaValidator,
+        InstrumentationDecorator, Partitioner, RetentionValidator, RpcWrite,
     },
     gossip::{
         namespace_cache::NamespaceSchemaGossip, schema_change_observer::SchemaChangeObserver,
@@ -63,6 +63,7 @@ use router::{
     namespace_resolver::{
         MissingNamespaceAction, NamespaceAutocreation, NamespaceResolver, NamespaceSchemaResolver,
     },
+    schema_validator::SchemaValidator,
     server::{
         grpc::RpcWriteGrpcDelegate,
         http::{
